Grammar/translation fix
Changed the line

"@When the world is changed to blue,{pause(15)} happiness can be established to the country." (which is bad grammar)

to

"@When the world is changed to blue,{pause(15)} a country of happiness will be established."

The original English translation had the line as "@When the world is changed to blue, a peaceful country can be established."
The Japanese text is "世界をブルーにかえたら幸せの国はやってくる。" which translates to "If you change the world to blue, a country of happiness will come." Since the line is promoting Happy Happyism, using the word "happiness" is more appropriate than "peaceful". (And is closer to the meaning of the Japanese word "幸せ".)

Keeping the word "established" from the original English translation is fine, changing "can" to "will" is mostly just a stylistic choice, implying that the speaker thinks this outcome is inevitable. (Reinforced by the use of "when" rather than "if".)